backend/requirements.txt
Flask==2.1.1
Flask-Cors==5.0.1
flask-mysqldb==1.0.1